third_party/blink/web_tests/platform/win11-arm64/external/wpt/webxr/plane-detection/xrSession_initiateRoomCapture.https-expected.txt

This is a testharness.js-based test.
[FAIL] initiateRoomCapture behavior - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] initiateRoomCapture behavior - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] initiateRoomCapture throws NotSupportedError if feature not requested - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] initiateRoomCapture throws NotSupportedError if feature not requested - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] initiateRoomCapture throws InvalidStateError if session ended - webgl
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
[FAIL] initiateRoomCapture throws InvalidStateError if session ended - webgl2
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'makeXRCompatible')"
Harness: the test ran to completion.
